+++ index.js
@@ -310,7 +310,7 @@ if (!nativeBinding) {
   throw new Error(`Failed to load native binding`)
 }
 
-const { JsClient, JsPostgresProvider, JsSqliteProvider, JsRuntime, activityTraceLog, orchestrationTraceLog, activityIsCancelled } = nativeBinding
+const { JsClient, JsPostgresProvider, JsSqliteProvider, JsRuntime, activityTraceLog, orchestrationTraceLog, activityIsCancelled, activityGetClient } = nativeBinding
 
 module.exports.JsClient = JsClient
 module.exports.JsPostgresProvider = JsPostgresProvider
@@ -319,3 +319,4 @@ module.exports.JsRuntime = JsRuntime
 module.exports.activityTraceLog = activityTraceLog
 module.exports.orchestrationTraceLog = orchestrationTraceLog
 module.exports.activityIsCancelled = activityIsCancelled
+module.exports.activityGetClient = activityGetClient